+++ docs/CHANGES.md
@@ -18,6 +18,9 @@ should be able to easily backport future new features to older versions rather e
 
 ## CHANGE LOG ##
 
+* @dev
+  * [GH-263] Fixed `Validator::assertIsType()` not validating provided exception class against `InvalidTypeExceptionContract`.
+
 * v12.0.1 (2025-04-16)
   * [RB-255] Fixed `ToArrayConverter` using a new Request instance instead of the actual request.
   * [BR-256] Fixed `ToArrayConverter` potentially causing fatal error if object lacks `toArray` method.